+++ app.go
@@ -2,6 +2,8 @@ package main
 
 import (
 	"context"
+	"crypto/tls"
+	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -33,6 +35,15 @@ func initCache() {
 	config.RedisClient = client
 }
 
+// convert Base64 Encodes PEM Certificate to tls object
+func decodeBase64Cert(pemCert string) []byte {
+	sslPem, err := base64.URLEncoding.DecodeString(pemCert)
+	if err != nil {
+		logrus.Fatal("Error decoding SSL PEM from Base64: ", err.Error())
+	}
+	return sslPem
+}
+
 func main() {
 	util.SetLogging(config.LogLevel, config.EnableSyslog, config.AppName)
 	logrus.Println(config.AppName + " build " + MinVersion)
@@ -60,9 +71,6 @@ func main() {
 	//		{Type: mesosproto.FrameworkInfo_Capability_RESERVATION_REFINEMENT},
 	//	}
 
-	// The Hostname should ever be set after reading the state file.
-	framework.FrameworkInfo.Hostname = &framework.FrameworkHostname
-
 	initCache()
 	mesosutil.SetConfig(&framework)
 	api.SetConfig(&config, &framework)
@@ -74,16 +82,35 @@ func main() {
 		json.Unmarshal([]byte(key), &framework)
 	}
 
-	// load framework config from DB
-	key = api.GetRedisKey("framework_config")
-	if key != "" {
-		json.Unmarshal([]byte(key), &config)
+	// The Hostname should ever be set after reading the state file.
+	framework.FrameworkInfo.Hostname = &framework.FrameworkHostname
+
+	server := &http.Server{
+		Addr:    listen,
+		Handler: api.Commands(),
+		TLSConfig: &tls.Config{
+			ClientAuth: tls.RequestClientCert,
+			MinVersion: tls.VersionTLS12,
+		},
 	}
 
-	http.Handle("/", api.Commands())
+	if config.SSLCrt != "" && config.SSLKey != "" {
+		logrus.Debug("Enable TLS")
+		crt := decodeBase64Cert(config.SSLCrt)
+		key := decodeBase64Cert(config.SSLKey)
+		certs, err := tls.X509KeyPair(crt, key)
+		if err != nil {
+			logrus.Fatal("TLS Server Error: ", err.Error())
+		}
+		server.TLSConfig.Certificates = []tls.Certificate{certs}
+	}
 
 	go func() {
-		http.ListenAndServe(listen, nil)
+		if config.SSLCrt != "" && config.SSLKey != "" {
+			server.ListenAndServeTLS("", "")
+		} else {
+			server.ListenAndServe()
+		}
 	}()
 	logrus.Fatal(mesos.Subscribe())
 }